+++ app/uk/gov/hmrc/eusubsidycompliancefrontend/util/TimeProvider.scala
@@ -18,14 +18,16 @@ package uk.gov.hmrc.eusubsidycompliancefrontend.util
 
 import com.google.inject.{ImplementedBy, Singleton}
 
-import java.time.LocalDate
+import java.time.{LocalDate, ZoneId}
 
-@ImplementedBy(classOf[TimeProviderImpl])
+@ImplementedBy(classOf[SystemTimeProvider])
 trait TimeProvider {
   def today: LocalDate
+  def today(z: ZoneId): LocalDate
 }
 
 @Singleton
-class TimeProviderImpl extends TimeProvider {
+class SystemTimeProvider extends TimeProvider {
   override def today: LocalDate = LocalDate.now()
+  override def today(z: ZoneId): LocalDate = LocalDate.now(z)
 }
